% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limonadeST.R
\name{limonadeST}
\alias{limonadeST}
\title{limonadeST}
\usage{
limonadeST(filename = NULL, format = NULL, limonadeST_param = NULL)
}
\arguments{
\item{filename}{filename or vector of filenames}

\item{format}{format of the file}

\item{limonadeST_param}{limonadeST_param object}
}
\description{
limonadeST
}
\examples{
\dontrun{
limonadeST()
}
}
\keyword{limonadeST}

